Plug1n Nightlies through Travis
Make travis ready to build nightlies once new pushes come in to master
This commit is made up of multiple commits that have been squashed,
more about those below.

Squashed commits:

[788f03d] Fix QR Code not showing up

The previous commit changes git's behavior to clear the /build folder,
so the QR Code needs to be imported from the main branch and put back in
on building.

[d73f150] Fix git not merging unrelated histories

The argument `--allow-unrelated-histories` needs to be added so master
can be merged into gh-branches and be built

[a1beeb7] Add new nightly location to Readme

Change nightly location from Homebr3w to Plug1n's Nightlies

[4058865] Make Plug1n ready for nightly building

Change the build script just ever so slightly to make it fit for nightly
building with Travis